Rewrite parse_archive as a filter
As fun as this utility was written in vimscript, it really wasn't very useful
because you could only handle one file at a time and you would end up in vim
(where you'd have to write and quit). This rewrites the utility as a filter and
even improves the output so that it contains thread breaks and the subjects.